+++ decomposition/omni_chassis/src/omni_chassis.cpp
@@ -60,8 +60,8 @@ private:
 
     std::unique_ptr<OmniKinematics> kine;
 
-    float gimbal_yaw_pos = 0.0; // The yaw position of the gimbal against the ground, in radians.
-    float motor_yaw_pos = 0.0; // The yaw position of the gimbal against the chassis, in radians.
+    double gimbal_yaw_pos = 0.0; // The yaw position of the gimbal against the ground, in radians.
+    double motor_yaw_pos = 0.0; // The yaw position of the gimbal against the chassis, in radians.
 
     void cha_callback(const behavior_interface::msg::Move::SharedPtr cha_msg)
     {
